Refactor todo agent tool definitions and implementations
- Moved CRUD tool implementations to separate modules: todo_crud_tools and todo_schedule_tools.
- Introduced helper functions to build tool objects and categorize them into universal, CRUD, schedule, and support tools.
- Updated get_tool_definitions to return a comprehensive list of tools.
- Removed the tool_implementations.py file as its contents were refactored into the new structure.
- Enhanced error handling and argument preprocessing in CRUD operations.
