+++ src/app/domain/todo_agents/tools/__init__.py
@@ -4,12 +4,30 @@ This module contains all the tool-related components for todo agents,
 organized into focused modules for better maintainability.
 """
 
-from .agent_factory import get_todo_agent
-from .tool_definitions import get_tool_definitions
+from .agent_factory import (
+    get_agent_by_name,
+    get_todo_agent,
+    get_todo_crud_agent,
+    get_todo_schedule_agent,
+    get_todo_support_agent,
+)
+from .tool_definitions import (
+    get_crud_tool_definitions,
+    get_schedule_tool_definitions,
+    get_support_tool_definitions,
+    get_tool_definitions,
+)
 from .universal_tools import get_user_datetime_impl
 
 __all__ = [
     "get_todo_agent",
+    "get_todo_crud_agent",
+    "get_todo_schedule_agent",
+    "get_todo_support_agent",
+    "get_agent_by_name",
     "get_tool_definitions",
+    "get_crud_tool_definitions",
+    "get_schedule_tool_definitions",
+    "get_support_tool_definitions",
     "get_user_datetime_impl",
 ]
